+++ src/index.ts
@@ -1,17 +1,1 @@
-import {
-  CANVAS,
-  CONTRACT_VERSIONS,
-  RENDER_CRAFT,
-  SCHEMA_VERSIONS,
-  VOCAB_V1,
-  beatRole,
-} from "./contracts/index.js";
-
-export {
-  CANVAS,
-  CONTRACT_VERSIONS,
-  RENDER_CRAFT,
-  SCHEMA_VERSIONS,
-  VOCAB_V1,
-  beatRole,
-};
+export * from "./contracts/index.js";